// src/AdminApp/outback-admin-old/src/app/services/generated/api.generated.ts
export enum ClientType {
    _0 = 0,
    _1 = 1,
}

export class CreateClient implements ICreateClient {
    clientName!: string;
    description!: string;
    clientType!: ClientType;
    familyId!: number;

    constructor(data?: ICreateClient) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.clientName = _data["clientName"];
            this.description = _data["description"];
            this.clientType = _data["clientType"];
            this.familyId = _data["familyId"];
        }
    }

    static fromJS(data: any): CreateClient {
        data = typeof data === 'object' ? data : {};
        let result = new CreateClient();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["clientName"] = this.clientName;
        data["description"] = this.description;
        data["clientType"] = this.clientType;
        data["familyId"] = this.familyId;
        return data;
    }
}

export interface ICreateClient {
    clientName: string;
    description: string;
    clientType: ClientType;
    familyId: number;
}
// ...
export class OutbackClientFamily implements IOutbackClientFamily {
    id?: number;
    name?: string;
    description?: string;
    created?: Date;
    updated?: Date;
    deleted?: Date | undefined;

    constructor(data?: IOutbackClientFamily) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.description = _data["description"];
            this.created = _data["created"] ? new Date(_data["created"].toString()) : <any>undefined;
            this.updated = _data["updated"] ? new Date(_data["updated"].toString()) : <any>undefined;
            this.deleted = _data["deleted"] ? new Date(_data["deleted"].toString()) : <any>undefined;
        }
    }

    static fromJS(data: any): OutbackClientFamily {
        data = typeof data === 'object' ? data : {};
        let result = new OutbackClientFamily();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["description"] = this.description;
        data["created"] = this.created ? this.created.toISOString() : <any>undefined;
        data["updated"] = this.updated ? this.updated.toISOString() : <any>undefined;
        data["deleted"] = this.deleted ? this.deleted.toISOString() : <any>undefined;
        return data;
    }
}

export interface IOutbackClientFamily {
    id?: number;
    name?: string;
    description?: string;
    created?: Date;
    updated?: Date;
    deleted?: Date | undefined;
}